// lib/api/generated/models/exerciseRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { ExerciseRecordKind } from './exerciseRecordKind';
import type { ExerciseRecordSource } from './exerciseRecordSource';
import type { ExerciseRecordStatus } from './exerciseRecordStatus';

export interface ExerciseRecord {
  achieved_on: string;
  exercise_id: string;
  id: string;
  kg: number;
  kind?: ExerciseRecordKind;
  /** @nullable */
  notes?: string | null;
  /** @minimum 0 */
  reps?: number;
  /** @nullable */
  reviewer_note?: string | null;
  source?: ExerciseRecordSource;
  status?: ExerciseRecordStatus;
  submitted_at: string;
  updated_at: string;
  user_id: string;
}