// lib/api/generated/models/exerciseRecordKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type ExerciseRecordKind = typeof ExerciseRecordKind[keyof typeof ExerciseRecordKind];


export const ExerciseRecordKind = {
  training: 'training',
  test: 'test',
} as const;